small fixes & merged {top_provider,top_stats}.hpp into top.hpp
