+++ monitor/monitor.h
@@ -106,7 +106,7 @@ void dnet_monitor_remove_provider(struct dnet_node *n, const char *name);
  * \a size - size of data that takes a part in command execution
  * \a time - time spended on command execution
  */
-void dnet_monitor_stats_update(struct dnet_node *n, const struct dnet_cmd *cmd, const int trans,
+void dnet_monitor_stats_update(struct dnet_node *n, const struct dnet_cmd *cmd,
                                const int err, const int cache,
                                const uint32_t size, const unsigned long time);
